requirements.txt
streamlit
requests
openpyxl
pillow
rapidocr_onnxruntime
numpy
onnxruntime